+++ azdev/help.py
@@ -135,6 +135,25 @@ helps['linter'] = """
           text: azdev linter --repo azure-cli --tgt upstream/master --src upstream/dev
 """
 
+helps['statistics'] = """
+    short-summary: Commands for CLI modules statistics.
+"""
+
+helps['statistics list-command-table'] = """
+    short-summary: List Command table for CLI module.
+    examples:
+        - name: List command table for the network module
+          text: azdev statistics list-command-table network -o table
+        - name: List command table for all modules of azure-cli repo, without commands details
+          text: azdev statistics list-command-table CLI --statistics-only
+"""
+
+helps['statistics diff-command-tables'] = """
+    short-summary: Diff the command table change.
+    examples:
+        - name: Diff the command table change from May to Oct
+          text: azdev statistics diff-command-tables --table-path command-table_May_01.json --diff-table-path command-table_Oct_01.json --statistics-only
+"""
 
 helps['perf'] = """
     short-summary: Commands to test CLI performance.