+++ azdev/operations/testtool/pytest_runner.py
@@ -18,7 +18,7 @@ def get_test_runner(parallel, log_path, last_failed, no_exit_first, mark):
         logger = get_logger(__name__)
 
         if os.name == 'posix':
-            arguments = ['-x', '-v', '--boxed', '-p no:warnings', '--log-level=WARN', '--junit-xml', log_path]
+            arguments = ['-x', '-v', '--forked', '-p no:warnings', '--log-level=WARN', '--junit-xml', log_path]
         else:
             arguments = ['-x', '-v', '-p no:warnings', '--log-level=WARN', '--junit-xml', log_path]